fix: review img put 수정... (추가 수정 필요)

+++ back/src/middlewares/file_delete.js
@@ -2,20 +2,18 @@ const fs = require("fs");
 const path = require("path");
 const maria = require("../db/connect/maria");
 
-const fileDelete = reviewId => {
+const fileDelete = async (reviewId, newImgName) => {
   maria.query(`SELECT reviewImg FROM REVIEW WHERE reviewId = ?`, [reviewId], async function (err, rows, fields) {
     if (err) {
       return res.send(err);
     }
 
-    const imgSrc = rows[0].reviewImg;
-    const imgName = imgSrc.replace(hostURL, "");
+    const imgSrc = await rows[0].reviewImg;
+    const imgName = await imgSrc.replace(hostURL, "");
 
-    if (imgName !== "default.jpg") {
+    if (imgName !== "default.jpg" && newImgName) {
       const imgPath = path.resolve(__dirname, "../../uploads", imgName);
       fs.unlinkSync(imgPath);
-    } else {
-      return;
     }
   });
 };
